requirements.txt
streamlit>=1.32
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
openai>=1.0
langgraph>=0.2.0
langchain-core>=0.3.0
tenacity>=8.2
pytest>=8.0
ruff>=0.9
